+++ packages/vanchor/src/wrappingFee/index.ts
@@ -1,0 +1,3 @@
+export { default as recordWrappingFee } from './wrappingFees';
+export { default as recordWrappingFee15MinsInterval } from './wrappingFees15MinsInterval';
+export { default as recordWrappingFeeDayInterval } from './wrappingFeesDayInterval';
