--- packages/vanchor/src/wrappingFees/index.ts
@@ -1,3 +1,0 @@
-export { default as recordWrappingFees } from './wrappingFees';
-export { default as recordWrappingFees15MinsInterval } from './wrappingFees15MinsInterval';
-export { default as recordWrappingFeesDayInterval } from './wrappingFeesDayInterval';
